# docs/openapi-rest.yaml
openapi: 3.0.3
info:
  title: LCP Platform API
  version: 1.1.0
  description: >
    REST and GraphQL surface for the self-hosted LCP platform. This document
    covers the admin dashboard login flow, publication processing, license
    lookup, status APIs, content downloads, publication catalog operations,
    runtime endpoints, and the GraphQL operations used by integration clients
    such as WordPress.
servers:
  - url: https://testmedical.ir
  - url: https://staging.testmedical.ir
security:
  - bearerAuth: []
paths:
  /api/v1/auth/login:
    post:
      summary: Sign in and mint a JWT for the admin dashboard or publisher flow
      operationId: authLogin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: JWT token response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/auth/ping:
    get:
      summary: Lightweight auth health check
      operationId: authPing
      security: []
      responses:
        "200":
          description: Auth subsystem is available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PingResponse"
  /api/v1/publications:
    get:
      summary: List publication catalog entries
      operationId: listPublications
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Publication catalog
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicationListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      summary: Create a publication record and optionally process a file
      operationId: createPublication
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublicationRequest"
      responses:
        "201":
          description: Publication created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Publication"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/publications/{publicationId}:
    get:
      summary: Get a publication record by ID
      operationId: getPublication
      security:
        - bearerAuth: []
      parameters:
        - name: publicationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Publication record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Publication"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      summary: Update publication metadata
      operationId: patchPublication
      security:
        - bearerAuth: []
      parameters:
        - name: publicationId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublicationPatchRequest"
      responses:
        "200":
          description: Publication updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Publication"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/publications/{publicationId}/activate:
    post:
      summary: Mark a publication active
      operationId: activatePublication
      security:
        - bearerAuth: []
      parameters:
        - name: publicationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Publication activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Publication"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/publications/{publicationId}/deactivate:
    post:
      summary: Mark a publication inactive
      operationId: deactivatePublication
      security:
        - bearerAuth: []
      parameters:
        - name: publicationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Publication deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Publication"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/lcp/process:
    post:
      summary: Process publication input data or an uploaded publication
      operationId: processPublication
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcessRequest"
      responses:
        "200":
          description: Processing completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessStatus"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/lcp/status:
    get:
      summary: Get service or process status
      operationId: getProcessStatus
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Status payload
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/StatusEnvelope"
                  - $ref: "#/components/schemas/ProcessStatus"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/admin/metrics:
    get:
      summary: Get admin metrics
      operationId: getAdminMetrics
      security:
        - bearerAuth: []
      parameters:
        - name: X-2FA-Code
          in: header
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Metrics payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetricsResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/admin/users:
    get:
      summary: List admin-visible users
      operationId: listAdminUsers
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUsersResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/admin/users/{userId}/verify:
    post:
      summary: Mark a user as verified
      operationId: verifyAdminUser
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/admin/users/{userId}/unverify:
    post:
      summary: Mark a user as unverified
      operationId: unverifyAdminUser
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/licenses/{licenseId}/user:
    get:
      summary: Get public user data for a license
      operationId: getLicenseUserData
      security: []
      parameters:
        - name: licenseId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LicenseUserData"
        "404":
          $ref: "#/components/responses/NotFound"
  /publications/{publicationId}/content:
    get:
      summary: Download encrypted publication content
      operationId: downloadPublicationContent
      security:
        - bearerAuth: []
      parameters:
        - name: publicationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Encrypted publication asset
          content:
            application/octet-stream: {}
        "404":
          $ref: "#/components/responses/NotFound"
  /graphql:
    post:
      summary: GraphQL endpoint for publication and license workflows
      operationId: graphQLOperation
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GraphQLRequest"
      responses:
        "200":
          description: GraphQL response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GraphQLResponse"
  /healthz:
    get:
      summary: Liveness probe
      operationId: healthz
      security: []
      responses:
        "200":
          description: Alive
  /readyz:
    get:
      summary: Readiness probe
      operationId: readyz
      security: []
      responses:
        "200":
          description: Ready
  /metrics:
    get:
      summary: Prometheus metrics
      operationId: metrics
      security: []
      responses:
        "200":
          description: Prometheus text exposition metrics
          content:
            text/plain:
              schema:
                type: string
  /swagger.yaml:
    get:
      summary: OpenAPI document in YAML format
      operationId: swaggerYaml
      security: []
      responses:
        "200":
          description: OpenAPI YAML
          content:
            text/yaml:
              schema:
                type: string
  /swagger.json:
    get:
      summary: OpenAPI document in JSON format
      operationId: swaggerJson
      security: []
      responses:
        "200":
          description: OpenAPI JSON
          content:
            application/json:
              schema:
                type: object
  /docs/openapi.yaml:
    get:
      summary: OpenAPI document in YAML format
      operationId: docsOpenapiYaml
      security: []
      responses:
        "200":
          description: OpenAPI YAML
          content:
            text/yaml:
              schema:
                type: string
  /docs/swagger.json:
    get:
      summary: OpenAPI document in JSON format
      operationId: docsSwaggerJson
      security: []
      responses:
        "200":
          description: OpenAPI JSON
          content:
            application/json:
              schema:
                type: object
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Invalid request payload
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Role or 2FA validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      additionalProperties: false
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: amiradmin
        password:
          type: string
          example: e40949a95921fb6e
        twoFactor:
          type: string
          example: 65b5ec
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        role:
          type: string
          example: admin
        subject:
          type: string
          example: amiradmin
        expiresAt:
          type: string
          format: date-time
    Publication:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        authors:
          type: array
          items:
            type: string
        language:
          type: string
        subjects:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        status:
          type: string
          example: active
        file_path:
          type: string
        encrypted_path:
          type: string
        encrypted_uri:
          type: string
        checksum:
          type: string
        licenseDurationDays:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PublicationListResponse:
      type: object
      properties:
        publications:
          type: array
          items:
            $ref: "#/components/schemas/Publication"
    PublicationRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
        authors:
          type: array
          items:
            type: string
        language:
          type: string
        subjects:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        encrypted_uri:
          type: string
        checksum:
          type: string
        license_duration_days:
          type: integer
        file:
          type: string
          description: Base64 encoded content of the uploaded publication
        status:
          type: string
          description: Publication status for catalog management
    PublicationPatchRequest:
      type: object
      properties:
        title:
          type: string
        authors:
          type: array
          items:
            type: string
        language:
          type: string
        subjects:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        encrypted_uri:
          type: string
        checksum:
          type: string
        license_duration_days:
          type: integer
        status:
          type: string
    PingResponse:
      type: object
      properties:
        status:
          type: string
          example: ok
    ProcessRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          example: Example Publication
        file:
          type: string
          description: Base64 encoded publication content selected in the browser
        data:
          type: object
          description: Optional opaque metadata payload for future extensions
    ProcessStatus:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
          example: completed
        publicationId:
          type: string
        error:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    StatusEnvelope:
      type: object
      properties:
        status:
          type: string
          example: ok
        uptimeSec:
          type: integer
          format: int64
        processes:
          type: array
          items:
            $ref: "#/components/schemas/ProcessStatus"
    MetricsResponse:
      type: object
      properties:
        uptimeSec:
          type: integer
          format: int64
        processes:
          type: integer
        metrics:
          type: object
          properties:
            requestsTotal:
              type: integer
            processesOk:
              type: integer
            processesFail:
              type: integer
    AdminUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
        verified:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AdminUsersResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AdminUser"
    LicenseUserData:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        passphraseHash:
          type: string
        hint:
          type: string
    GraphQLRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          description: >
            Supported operations include uploadPublication, createLicense,
            revokeLicense, licenses, and publications.
        variables:
          type: object
          additionalProperties: true
    GraphQLResponse:
      type: object
      properties:
        data:
          type: object
          additionalProperties: true
        errors:
          type: array
          items:
            type: object
            properties:
              message:
                type: string